requirements.txt
asciichartpy==1.5.7
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
pandas==0.24.1
